requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.0.1=pypi_0
anyio=3.5.0=py38hecd8cb5_0
appnope=0.1.2=py38hecd8cb5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38hca72f7f_0
asn1crypto=1.5.1=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
backports-zoneinfo=0.2.1=pypi_0
beautifulsoup4=4.12.2=py38hecd8cb5_0
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.6.2=pypi_0
ca-certificates=2023.05.30=hecd8cb5_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=py38h6c40b1e_3
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.0.0=pypi_0
comm=0.1.2=py38hecd8cb5_0
coverage=7.2.7=pypi_0
cryptography=40.0.2=pypi_0
dataclasses-json=0.5.8=pypi_0
debugpy=1.5.1=py38he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
faiss-cpu=1.7.4=pypi_0
filelock=3.12.2=pypi_0
frozenlist=1.3.3=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
google-auth=2.20.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
greenlet=2.0.2=pypi_0
gsheetsdb=0.1.13.1=pypi_0
gspread=5.9.0=pypi_0
icu=58.2=h0a44026_3
idna=3.4=py38hecd8cb5_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
importlib_metadata=6.0.0=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
ipykernel=6.19.2=py38h01d92e1_0
ipython=8.12.0=py38hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jaraco-classes=3.2.3=pypi_0
jedi=0.18.1=py38hecd8cb5_1
jinja2=3.1.2=py38hecd8cb5_0
jsonschema=4.17.3=py38hecd8cb5_0
jupyter_client=8.1.0=py38hecd8cb5_0
jupyter_core=5.3.0=py38hecd8cb5_0
jupyter_events=0.6.3=py38hecd8cb5_0
jupyter_server=2.5.0=py38hecd8cb5_0
jupyter_server_terminals=0.4.4=py38hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
keyring=23.13.1=pypi_0
langchain=0.0.199=pypi_0
langchainplus-sdk=0.0.9=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
libiconv=1.16=hca72f7f_2
libsodium=1.0.18=h1de35cc_0
libxml2=2.10.3=h930c0e2_0
libxslt=1.1.37=h6d1eb0e_0
lxml=4.9.2=py38h6c40b1e_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.19.0=pypi_0
marshmallow-enum=1.5.1=pypi_0
matplotlib-inline=0.1.6=py38hecd8cb5_0
mdurl=0.1.2=pypi_0
mistune=0.8.4=py38h1de35cc_1001
mo-dots=4.22.21108=pypi_0
mo-future=3.147.20327=pypi_0
mo-imports=3.149.20327=pypi_0
mo-kwargs=4.22.21108=pypi_0
mo-logs=4.23.21108=pypi_0
more-itertools=9.1.0=pypi_0
moz-sql-parser=4.40.21126=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclassic=0.5.5=py38hecd8cb5_0
nbclient=0.5.13=py38hecd8cb5_0
nbconvert=6.5.4=py38hecd8cb5_0
nbformat=5.7.0=py38hecd8cb5_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.5.6=py38hecd8cb5_0
notebook=6.5.4=py38hecd8cb5_0
notebook-shim=0.2.2=py38hecd8cb5_0
numexpr=2.8.4=pypi_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
openai=0.27.8=pypi_0
openapi-schema-pydantic=1.2.4=pypi_0
openssl=3.0.8=hca72f7f_0
oscrypto=1.3.0=pypi_0
packaging=23.1=pypi_0
pandas=2.0.2=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.5.0=pypi_0
pip=23.1.2=py38hecd8cb5_0
pkgutil-resolve-name=1.3.10=py38hecd8cb5_0
platformdirs=2.5.2=py38hecd8cb5_0
prometheus_client=0.14.1=py38hecd8cb5_0
prompt-toolkit=3.0.36=py38hecd8cb5_0
protobuf=4.23.2=pypi_0
psutil=5.9.0=py38hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=12.0.1=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodomex=3.18.0=pypi_0
pydantic=1.10.9=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.15.1=py38hecd8cb5_1
pyjwt=2.7.0=pypi_0
pympler=1.0.1=pypi_0
pyopenssl=23.2.0=pypi_0
pyparsing=3.1.0=pypi_0
pypdf2=3.0.1=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.8.16=h5ee71fb_4
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.0=pypi_0
python-fastjsonschema=2.16.2=py38hecd8cb5_0
python-json-logger=2.0.7=py38hecd8cb5_0
pytz=2023.3=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=6.0=py38h6c40b1e_1
pyzmq=25.1.0=py38hcec6c5f_0
readline=8.2=hca72f7f_0
regex=2023.6.3=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
requirements-parser=0.5.0=pypi_0
rfc3339-validator=0.1.4=py38hecd8cb5_0
rfc3986-validator=0.1.1=py38hecd8cb5_0
rich=13.4.2=pypi_0
rsa=4.9=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=67.8.0=py38hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.2.0=py38hecd8cb5_1
snowflake-cli-labs=0.2.9=pypi_0
snowflake-connector-python=3.0.4=pypi_0
snowflake-snowpark-python=1.4.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.4=py38hecd8cb5_0
sqlalchemy=2.0.16=pypi_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.23.1=pypi_0
streamlit-ace=0.1.1=pypi_0
streamlit-chat=0.0.2.2=pypi_0
tabulate=0.9.0=pypi_0
terminado=0.17.1=py38hecd8cb5_0
tiktoken=0.4.0=pypi_0
tinycss2=1.2.1=py38hecd8cb5_0
tk=8.6.12=h5d9f67b_0
toolz=0.12.0=pypi_0
tornado=6.3.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py38hecd8cb5_0
typer=0.9.0=pypi_0
types-setuptools=67.8.0.0=pypi_0
typing-extensions=4.6.3=py38hecd8cb5_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.6.3=py38hecd8cb5_0
tzdata=2023.3=pypi_0
tzlocal=4.3=pypi_0
urllib3=1.26.16=pypi_0
validators=0.20.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38hecd8cb5_4
wheel=0.38.4=py38hecd8cb5_0
xz=5.4.2=h6c40b1e_0
yaml=0.2.5=haf1e3a3_0
yarl=1.9.2=pypi_0
zeromq=4.3.4=h23ab428_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h4dc903c_0
